requirements.txt
python-dotenv
openrgb-python
python-socketio[client]
